- Added logs command and its tests - Added tests for member start with config file - Added processUniqueId to HazelcastProcess - Minor improvements

+++ src/main/java/com/hazelcast/commandline/MemberCommandLine.java
@@ -4,16 +4,16 @@ import com.hazelcast.core.HazelcastException;
 
 import java.io.*;
 import java.lang.reflect.Field;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.nio.file.Paths;
+import java.util.*;
 import java.util.concurrent.Callable;
 import java.util.stream.Stream;
 
+import static com.hazelcast.commandline.HazelcastCommandLine.*;
 import static picocli.CommandLine.*;
 
 @Command(
@@ -26,12 +26,23 @@ import static picocli.CommandLine.*;
 public class MemberCommandLine implements Callable<Void> {
     private final PrintStream out;
     private final PrintStream err;
-    private final String FILE_PATH = "instances.dat";
+    private final String logsDirString = "logs";
+    private final String logsFileNameString = "hazelcast.log";
     private Stream<String> processOutput;
+    public final String instancesFilePath = HAZELCAST_HOME + SEPARATOR + "instances.dat";
 
     public MemberCommandLine(PrintStream out, PrintStream err) {
         this.out = out;
         this.err = err;
+        createHomeDirectory();
+    }
+
+    private void createHomeDirectory() {
+        try {
+            new File(HAZELCAST_HOME).mkdirs();
+        } catch (Exception e) {
+            throw new HazelcastException("Process directories couldn't created.");
+        }
     }
 
     public Void call() {
@@ -45,14 +56,51 @@ public class MemberCommandLine implements Callable<Void> {
             @Option(names = {"-c", "--config"},
             description = "Use <file> for Hazelcast configuration.")
             String configFilePath) throws IOException, ClassNotFoundException {
-        println("Starting a new Hazelcast IMDG member...");
-        List<String> parameters = new ArrayList<>();
+        List<String> args = new ArrayList<>();
         if (configFilePath != null && !configFilePath.isEmpty()) {
-            parameters.add("-Dhazelcast.config=" + configFilePath);
+            args.add("-Dhazelcast.config=" + configFilePath);
         }
-        Integer pid = buildJavaProcess(HazelcastMember.class, parameters);
-        saveProcess(new HazelcastProcess(pid));
-        println("Hazelcast IMDG Member started with pid: " + pid);
+
+        //TODO replace
+        String processUniqueId = UUID.randomUUID().toString();
+        String processDir = createProcessDirs(processUniqueId);
+        String loggingPropertiesPath = createLoggingPropertiesFile(processUniqueId, processDir);
+        args.add("-Djava.util.logging.config.file=" + loggingPropertiesPath);
+
+        Integer pid = buildJavaProcess(HazelcastMember.class, args);
+        saveProcess(new HazelcastProcess(processUniqueId, pid));
+
+        println(processUniqueId);
+    }
+
+    private String createProcessDirs(String processUniqueId) {
+        String processPath = HAZELCAST_HOME + SEPARATOR + processUniqueId;
+        String logPath = processPath + SEPARATOR + logsDirString;
+        try {
+            new File(logPath).mkdirs();
+        } catch (Exception e) {
+            throw new HazelcastException("Process directories couldn't created.");
+        }
+        return processPath;
+    }
+
+    private String createLoggingPropertiesFile(String processUniqueId, String processDir)
+            throws FileNotFoundException {
+        String loggingPropertiesPath = processDir + SEPARATOR + "logging.properties";
+        PrintWriter printWriter = new PrintWriter(loggingPropertiesPath);
+        String fileContent = "handlers= java.util.logging.FileHandler, java.util.logging.ConsoleHandler\n" +
+                ".level= INFO\n" +
+                "java.util.logging.FileHandler.pattern = " +
+                processDir + SEPARATOR + logsDirString + SEPARATOR + "hazelcast.log\n" +
+                "java.util.logging.FileHandler.limit = 50000\n" +
+                "java.util.logging.FileHandler.count = 1\n" +
+                "java.util.logging.FileHandler.maxLocks = 100\n" +
+                "java.util.logging.FileHandler.formatter = java.util.logging.SimpleFormatter\n" +
+                "java.util.logging.FileHandler.append=true\n" +
+                "java.util.logging.ConsoleHandler.formatter = java.util.logging.SimpleFormatter";
+        printWriter.println(fileContent);
+        printWriter.close();
+        return loggingPropertiesPath;
     }
 
     @Command(description = "Stops a Hazelcast IMDG member",
@@ -60,53 +108,80 @@ public class MemberCommandLine implements Callable<Void> {
     )
     public void stop(
             @Parameters(index = "0",
-            paramLabel = "<pid>",
+            paramLabel = "<process-id>",
             description = "Process id of the process to stop")
-                                 Integer pid) throws IOException, ClassNotFoundException {
-        println("Stopping Hazelcast IMDG member with pid: " + pid);
-
+                                 String processUniqueId) throws IOException, ClassNotFoundException {
+        HazelcastProcess process = getProcessMap().get(processUniqueId);
+        if (process == null){
+            printlnErr("No process found with process id: " + processUniqueId);
+            return;
+        }
+        int pid = process.getPid();
         Runtime.getRuntime().exec("kill -15 " + pid);
-        removeProcess(pid);
+        removeProcess(processUniqueId);
+        println(processUniqueId + " stopped.");
     }
 
     @Command(description = "Lists running Hazelcast IMDG members",
             mixinStandardHelpOptions = true
     )
     public void list() throws ClassNotFoundException {
-        Map<Integer, HazelcastProcess> processMap = getProcessMap();
+        Map<String, HazelcastProcess> processMap = getProcessMap();
         if (processMap.isEmpty()){
             println("No running process exists.");
             return;
         }
         for (HazelcastProcess process : processMap.values()) {
-            println(process.getPid() + "");
+            println(process.getProcessUniqueId());
         }
     }
 
-    private void removeProcess(Integer pid) throws IOException, ClassNotFoundException {
-        Map<Integer, HazelcastProcess> processMap = getProcessMap();
-        if (processMap == null || !processMap.containsKey(pid)){
-            println("No process found with pid: " + pid);
+    @Command(description = "Display the logs for Hazelcast member with the given ID.",
+            mixinStandardHelpOptions = true
+    )
+    public void logs(@Parameters(index = "0",
+            paramLabel = "<pid>",
+            description = "Process id of the process to show the logs")
+                                 String processUniqueId) throws ClassNotFoundException, IOException {
+        if (!getProcessMap().containsKey(processUniqueId)) {
+            printlnErr("No process found with process id: " + processUniqueId);
+        }
+        println(getLogs(processUniqueId));
+    }
+
+    private String getLogs(String processUniqueId) throws IOException {
+        String logsPath = HAZELCAST_HOME + SEPARATOR + processUniqueId + SEPARATOR
+                + logsDirString + SEPARATOR + logsFileNameString;
+        StringBuilder output = new StringBuilder();
+        Stream<String> stream = Files.lines( Paths.get(logsPath), StandardCharsets.UTF_8);
+        stream.forEach(s -> output.append(s).append("\n"));
+        return output.toString();
+    }
+
+    private void removeProcess(String processUniqueId) throws IOException, ClassNotFoundException {
+        Map<String, HazelcastProcess> processMap = getProcessMap();
+        if (processMap == null || !processMap.containsKey(processUniqueId)){
+            println("No process found with pid: " + processUniqueId);
             return;
         }
-        processMap.remove(pid);
+        processMap.remove(processUniqueId);
         updateFile(processMap);
     }
 
-    private Map<Integer, HazelcastProcess> getProcessMap() throws ClassNotFoundException {
+    protected Map<String, HazelcastProcess> getProcessMap() throws ClassNotFoundException {
         FileInputStream fileInputStream;
-        Map<Integer, HazelcastProcess> processes = new HashMap<>();
+        Map<String, HazelcastProcess> processes = new HashMap<>();
         try {
-            Path path = FileSystems.getDefault().getPath(FILE_PATH);
+            Path path = FileSystems.getDefault().getPath(instancesFilePath);
             if (!Files.exists(path)){
                 Files.createFile(path);
             }
-            fileInputStream = new FileInputStream(FILE_PATH);
+            fileInputStream = new FileInputStream(instancesFilePath);
             if (Files.size(path) == 0){
                 return processes;
             }
             ObjectInputStream input = new ObjectInputStream(fileInputStream);
-            processes = (Map<Integer, HazelcastProcess>) input.readObject();
+            processes = (Map<String, HazelcastProcess>) input.readObject();
             input.close();
         } catch (IOException e) {
             throw new HazelcastException("Error when reading from file.", e);
@@ -114,17 +189,17 @@ public class MemberCommandLine implements Callable<Void> {
         return processes;
     }
 
-    private Integer buildJavaProcess(Class aClass, List<String> parameters) throws IOException{
+    private Integer buildJavaProcess(Class aClass, List<String> parameters)
+            throws IOException{
         List<String> commandList = new ArrayList<>();
-        String separator = System.getProperty("file.separator");
         String classpath = System.getProperty("java.class.path");
         String path = System.getProperty("java.home")
-                + separator + "bin" + separator + "java";
+                + SEPARATOR + "bin" + SEPARATOR + "java";
         commandList.add(path);
         commandList.add("-cp");
         commandList.add(classpath);
-        commandList.add(aClass.getName());
         commandList.addAll(parameters);
+        commandList.add(aClass.getName());
         ProcessBuilder processBuilder = new ProcessBuilder(commandList);
         processBuilder.redirectErrorStream(true);
         Process process = processBuilder.start();
@@ -134,13 +209,13 @@ public class MemberCommandLine implements Callable<Void> {
     }
 
     private void saveProcess(HazelcastProcess process) throws IOException, ClassNotFoundException {
-        Map<Integer, HazelcastProcess> processMap = getProcessMap();
-        processMap.put(process.getPid(), process);
+        Map<String, HazelcastProcess> processMap = getProcessMap();
+        processMap.put(process.getProcessUniqueId(), process);
         updateFile(processMap);
     }
 
-    private void updateFile(Map<Integer, HazelcastProcess> processMap) throws IOException {
-        FileOutputStream fileOut = new FileOutputStream(FILE_PATH);
+    private void updateFile(Map<String, HazelcastProcess> processMap) throws IOException {
+        FileOutputStream fileOut = new FileOutputStream(instancesFilePath);
         ObjectOutputStream objectOut = new ObjectOutputStream(fileOut);
         objectOut.writeObject(processMap);
         objectOut.close();
@@ -171,6 +246,10 @@ public class MemberCommandLine implements Callable<Void> {
 
     private void println(String msg) {
         out.println(msg);
+    }
+
+    private void printlnErr(String msg) {
+        err.println(msg);
     }
 
     public Stream<String> getProcessOutput() {
